merubah susunan class dan activity untuk mengerjakan tugas pertemuan 4
